bronze/pyproject.toml
[project]
name = "ai-employee-bronze"
version = "0.1.0"
description = "Personal AI Employee - Bronze Tier: Foundation"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=6.0.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
ai-employee = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
